--- apps/main/src/dex/entities/gauge/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './reward-actions'
-export * from './gauge-info'
