+++ apps/main/src/dex/entities/gauge/model/gauge.query.ts
@@ -1,5 +1,5 @@
 import { type Address, zeroAddress } from 'viem'
-import { getGauge } from '@/dex/entities/gauge/lib'
+import { getGauge } from '@/dex/entities/gauge/lib/gauge-info'
 import { GaugeParams, GaugeQuery, rootKeys } from '@evm-ui/lib/model/query'
 import { queryFactory } from '@evm-ui/lib/model/query/factory'
 import { poolValidationSuite } from '@evm-ui/lib/model/query/pool-validation'
