--- apps/main/src/dex/entities/gauge/model/index.ts
@@ -1,3 +1,0 @@
-export * from './gauge.query'
-export * from './gauge-gas.query'
-export * from './gauge-validation'
